Update activity.c
In this code, the isPacketCorrupted function calculates the CRC (Cyclic Redundancy Check) for the packet data by summing up all the bytes in the data array. It then compares the calculated CRC with the value stored in the crc field of the packet. If the two values don't match, it indicates that the packet is corrupted.

You can call the isPacketCorrupted function for each received packet to check if it is corrupted before further processing or handling it accordingly.